phabert_cnn/requirements.txt
torch>=2.0.0
transformers>=4.30.0
triton>=2.0.0
einops>=0.6.0
datasets>=2.14.0
scikit-learn>=1.3.0
biopython>=1.81
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
imbalanced-learn>=0.11.0
